Move Cyclic focus logic to Modal component
